Task 2

+++ 0x1A-hash_tables/hash_tables.h
@@ -42,6 +42,6 @@ hash_table_t *hash_table_create(unsigned long int size);
 
 unsigned long int hash_djb2(const unsigned char *str);
 
-unsigned int key_index(const unsigned char *key, unsigned long int size);
+unsigned long int key_index(const unsigned char *key, unsigned long int size);
 
 #endif /* HASH_T */
